+++ lib/sqflite_migrator_plus.dart
@@ -2,3 +2,4 @@ library sqflite_migrator_plus;
 
 export './framework/migration_plus_class.dart';
 export './migrator_plus/db_migrator_initializer.dart';
+export './migrator_plus/db_migrator_runner.dart';