+++ CHANGELOG.md
@@ -1,3 +1,8 @@
 ## 0.0.1
 
-* TODO: Describe initial release.
+* This release implements the DbMigratorInitializer
+
+
+## 0.0.2
+
+* This release implements the DbMigratorRunner with migrations table